% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{aggregate_scores}
\alias{aggregate_scores}
\title{Aggregate scores of a GRanges object on a field}
\usage{
aggregate_scores(scored_granges, group_col, aggregate_by)
}
\arguments{
\item{scored_granges}{A GRanges object with numerical metadata columns}

\item{group_col}{A column among the mcols that can be seen as a factor.}

\item{aggregate_by}{Function used to aggregate: mean, median, true_mean.\preformatted{true_mean: Mean coverage taking all elements in a class as one large bin.

mean: mean-of-distribution approach. The mean of the aggregated value per
  locus is reported.

median: median-of-distribution. The median of the aggregated value per
  locus is reported.
}

FIXME: importFrom tidyselect where does not work according to:
https://github.com/r-lib/tidyselect/issues/201
This does not pass the checks but does not break the build process either
so I left it as is, so the note thrown by build is noted everytime.}
}
\value{
A data frame with aggregated scores.
}
\description{
Aggregates scores of a GRanges object on a specific field. Used for summary
functions.
}
