% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{bed_to_bw}
\alias{bed_to_bw}
\title{Parse a scored BED file and exports a bigWig file with given chromsizes}
\usage{
bed_to_bw(bed, bw, chromsizes)
}
\arguments{
\item{bed}{BED filename}

\item{bw}{bigWig filename}

\item{chromsizes}{Vector with length of chromosomes}
}
\description{
Parse a scored BED file and exports a bigWig file with given chromsizes
}
