% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{build_bins}
\alias{build_bins}
\title{Build a unscored bins GRanges object.}
\usage{
build_bins(bin_size = 10000, genome = "mm9", canonical = FALSE)
}
\arguments{
\item{bin_size}{Bin size.}

\item{genome}{Genome. Supported: any registered NCBI or UCSC genome
see registered_UCSC_genomes(), registered_NCBI_assemblies() from
GenomeInfoDb}

\item{canonical}{Use only canonical chromosomes (default: FALSE)}
}
\value{
A GRanges object with a tiled genome
}
\description{
Build a GRanges of bins of a given size, for a specific genome. Supported
genomes rely on \link[GenomeInfoDb]{Seqinfo}. This requires internet access
to work.
}
\examples{

build_bins(bin_size = 50000, genome = "mm9")
build_bins(bin_size = 50000, genome = "hg38")
build_bins(bin_size = 50000, genome = "mm10", canonical = TRUE)
}
