% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{build_bins_bw}
\alias{build_bins_bw}
\title{Build a tiles GRanges out of a bigWig file on a given bin size.}
\usage{
build_bins_bw(bwfile, bin_size)
}
\arguments{
\item{bwfile}{File to tile}

\item{bin_size}{Size of the tile in basepairs}
}
\value{
A GRanges object with bins of bin_size bins matching the genome the
bigWig file was mapped to.
}
\description{
It does not need the reference genome as that information can be parsed from
the bigwig file
}
\examples{

bw <- system.file("extdata", "sample_H33_ChIP.bw", package="wigglescout")
build_bins_bw(bw, 50000)
}
