% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_bed}
\alias{bw_bed}
\title{Build a bed-scored GRanges object from a bigWig file list and a BED file.}
\usage{
bw_bed(
  bwfiles,
  bedfile,
  bg_bwfiles = NULL,
  labels = NULL,
  per_locus_stat = "mean",
  aggregate_by = NULL,
  norm_mode = "fc",
  remove_top = 0
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bedfile}{BED file to intersect with the BigWig file.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{aggregate_by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}
}
\description{
Build a scored GRanges object from a bigWig file list and a BED file.
The aggregating function (per locus) can be min, max, sd, mean.
}
\details{
bwfiles and bg_bwfiles must have the same length. If you are using same
background for several files, then file paths must be repeated accordingly.

Values can be normalized using background bigWig files (usually input
files). By default, the value obtained will be bigwig / bg_bigwig per bin,
per bigWig.

If norm_func is specified, this can be changed to any given function, for
instance, if norm_func = log2, values will represent log2(bigwig / bg_bigwig)
per bin.
}
