% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_chr_coverage}
\alias{bw_chr_coverage}
\title{Calculate bigWig coverage per chromosome}
\usage{
bw_chr_coverage(bwfile, default_na = NA_real_)
}
\arguments{
\item{bwfile}{bigWig file}

\item{default_na}{Value to replace missing values.}
}
\description{
Calculate bigWig coverage per chromosome
}
