% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_global_coverage}
\alias{bw_global_coverage}
\title{Calculate bigWig global coverage.}
\usage{
bw_global_coverage(bwfile, default_na = NA_real_)
}
\arguments{
\item{bwfile}{bigWig file}

\item{default_na}{Value to replace missing values.}
}
\description{
Calculate bigWig global coverage.
}
