% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_heatmap}
\alias{bw_heatmap}
\title{Calculate heatmap matrix of a bigWig file over a BED file}
\usage{
bw_heatmap(
  bwfiles,
  bg_bwfiles = NULL,
  bedfile = NULL,
  labels = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  norm_func = identity
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{bedfile}{BED file to summarize}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{middle}{Number of base pairs that the middle section has (in stretch
mode). If not provided, median length of all loci is used.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{norm_func}{Function to apply to normalize bin values f(bw / bw_bg).}
}
\description{
Calculate heatmap matrix of a bigWig file over a BED file
}
