% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_loci}
\alias{bw_loci}
\title{Score a bigWig file list and a BED file or GRanges object.}
\usage{
bw_loci(
  bwfiles,
  loci,
  bg_bwfiles = NULL,
  labels = NULL,
  per_locus_stat = "mean",
  aggregate_by = NULL,
  norm_mode = "fc",
  remove_top = 0
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{loci}{GRanges or BED file to summarize the BigWig file at.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{aggregate_by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}
}
\value{
A GRanges object with each bwfile as a metadata column named
after labels, if provided, or after filenames otherwise.
}
\description{
Build a scored GRanges object from a bigWig file list and a BED file or
GRanges object. The aggregating function (per locus) can be min, max,
sd, mean.
}
\details{
Values can be normalized using background bigWig files (usually input
files). By default, the value obtained will be bigwig / bg_bigwig per locus,
per bigWig.

bwfiles and bg_bwfiles must have the same length. If you are using same
background for several files, then file paths must be repeated accordingly.

norm_mode can be either "fc", where values will represent bigwig / bg_bigwig,
or "log2fc", values will represent log2(bigwig / bg_bigwig) per locus.
}
\examples{
# Get the raw files
bed <- system.file("extdata", "sample_genes_mm9.bed", package="wigglescout")
bw <- system.file("extdata", "sample_H33_ChIP.bw", package="wigglescout")
bw2 <- system.file("extdata",
                   "sample_H3K9me3_ChIP.bw", package="wigglescout")
bw_inp <- system.file("extdata", "sample_Input.bw", package="wigglescout")

# Run single bw with single bed
bw_loci(bw, bed)

# Use of some parameters
bw_loci(bw, bed, labels = c("H33"), remove_top = 0.01)

# Log2 fold change
bw_loci(bw, loci = bed, bg_bwfiles = bw_inp, norm_mode = "log2fc")

# Multiple bigWig
bw_loci(c(bw, bw2),
        loci = bed,
        bg_bwfiles = c(bw_inp, bw_inp),
        norm_mode = "log2fc")
}
