% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_profile}
\alias{bw_profile}
\title{Calculate profile values of a bigWig file over a BED file.}
\usage{
bw_profile(
  bwfiles,
  bg_bwfiles = NULL,
  loci = NULL,
  labels = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  norm_mode = "fc",
  remove_top = 0
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{loci}{BED file or GRanges to summarize}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{middle}{Number of base pairs that the middle section has (in stretch
mode). If not provided, median length of all loci is used.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}
}
\value{
a data frame in long format
}
\description{
Calculates profile values of a set of tracks over the loci speficied in a
BED file. Data points are taken each bin_size base pairs.
}
\details{
Loci are aligned depending on mode parameter:
\itemize{
\item stretch. Aligns all starts and all ends, sort of stretching the loci.
The median of these lenghts is taken as the pseudo-length in order to show
a realistic plot when displayed.
\item start. All loci are aligned by start.
\item end. All loci are aligned by end.
\item center. All loci are aligned by center.
}
}
