% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_ranges}
\alias{bw_ranges}
\title{Score a GRanges object against a BigWig file}
\usage{
bw_ranges(bwfile, granges, per_locus_stat = "mean", selection = NULL)
}
\arguments{
\item{bwfile}{Path to a single BigWig file to be summarized.}

\item{granges}{GRanges object file to be summarized.}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}
}
\value{
GRanges with column score.
}
\description{
Build a scored GRanges object from a GRanges object and a single BigWig file.
The aggregating function can be min, max, sd, mean.
}
