% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{calculate_color_limits}
\alias{calculate_color_limits}
\title{Calculate color limits from a value matrix and provided parameters}
\usage{
calculate_color_limits(m, zmin, zmax)
}
\arguments{
\item{m}{Value matrix.}

\item{zmin}{Min value. Overrides percentile.}

\item{zmax}{Max value. Overrides percentile.}
}
\value{
A pair of c(min, max)
}
\description{
Calculate color limits from a value matrix and provided parameters
}
