% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{calculate_matrix_norm}
\alias{calculate_matrix_norm}
\title{Calculate a normalized heatmap matrix for a bigWig file over a BED file}
\usage{
calculate_matrix_norm(
  bw,
  granges,
  bg_bw = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  norm_func = identity,
  remove_top = 0
)
}
\arguments{
\item{bw}{BigWig file to be summarized.}

\item{granges}{GRanges object}

\item{bg_bw}{BigWig file to be used as background.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{middle}{Number of base pairs that the middle section has (in stretch
mode). If not provided, median length of all loci is used.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{norm_func}{Function to apply to normalize bin values f(bw / bw_bg).}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}
}
\description{
Calculate a normalized heatmap matrix for a bigWig file over a BED file
}
