% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{.calculate_stretch_matrix}
\alias{.calculate_stretch_matrix}
\title{Calculate matrix for stretch mode}
\usage{
.calculate_stretch_matrix(
  bw,
  granges,
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  default_na = NA_real_,
  scaling = "none"
)
}
\arguments{
\item{bw}{BigWigFile object}

\item{granges}{GRanges object}

\item{bin_size}{Bin size}

\item{upstream}{Number of basepairs upstream}

\item{downstream}{Number of basepairs downstream}

\item{middle}{Number of base pairs that the middle section has. If not
provided, median is used.}

\item{ignore_strand}{Ignore strand (bool)}

\item{default_na}{Default value for missing values}

\item{scaling}{none if no scaling is performed (default), relative if 1x
scaling is done (divide by global coverage).}
}
\value{
Summary matrix
}
\description{
Calculate matrix for stretch mode
}
