% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.colorscale}
\alias{.colorscale}
\title{Summary heatmap colorscale.}
\usage{
.colorscale(norm_mode, bg_bwfiles)
}
\arguments{
\item{norm_mode}{Type of normalization.}

\item{bg_bwfiles}{Background files.}
}
\value{
a ggproto object
}
\description{
Summary heatmap colorscale.
}
