% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.density_body}
\alias{.density_body}
\title{Density 2d histogram plot}
\usage{
.density_body(gr, binwidth = 0.05)
}
\arguments{
\item{gr}{GRanges}

\item{binwidth}{Size of the bin in histogram (resolution of the plot)}
}
\value{
A ggplot object
}
\description{
Plots a 2d histogramfrom a GRanges object that already has x and y columns.
}
