% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filter_violin_data}
\alias{.filter_violin_data}
\title{Filter violin plot data with quantile threshold}
\usage{
.filter_violin_data(gr, remove_top, columns)
}
\arguments{
\item{gr}{GRanges object}

\item{remove_top}{(1-remove_top) quantile will be removed.}

\item{columns}{mcols in GRanges to use for selection.}
}
\value{
A named list with ranges = GRanges object and stats a named list with
some calculated values.
}
\description{
Removes top quantile loci from the distribution. Values are removed by mean
rather than maximum.
}
