% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_labels}
\alias{.format_labels}
\title{Shorten labels to a max length and verify no repeated elements are produced}
\usage{
.format_labels(labels, max_length = 35)
}
\arguments{
\item{labels}{List of labels}

\item{max_length}{Number of characters the shorten labels can have (35)}
}
\value{
An array of strings
}
\description{
Shorten labels to a max length and verify no repeated elements are produced
}
