% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_wrapper_parameter_values}
\alias{.get_wrapper_parameter_values}
\title{Get matching parameters of a target function with current context}
\usage{
.get_wrapper_parameter_values(func, current)
}
\arguments{
\item{func}{Target function}

\item{current}{Current context as a named list}
}
\value{
Named list argument - value
}
\description{
This is a helper function for wrapping functions such as plotting. It
gets the formal arguments of the target function func, and gets their
values from the context given in current, that will have the same name.
It returns a named list argument - value that can be passed to func
using do.call()
}
