% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.granges_cbind}
\alias{.granges_cbind}
\title{GRanges cbind-like operation}
\usage{
.granges_cbind(grlist, labels)
}
\arguments{
\item{grlist}{A list of GRanges objects that have all the same fields.}

\item{labels}{Vector of names for the score columns.}
}
\description{
Perform a cbind operation on a GRanges list, appending scores to mcols.
It will sort the GRanges elements in order to ensure the match is proper.
}
\details{
This assumes that you are trying to cbind things that match (bins generated
from the same parameters, BED intersections from the same BED.)
}
