% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.heatmap_lines}
\alias{.heatmap_lines}
\title{Helper function for calculating guide lines and labels in heatmap}
\usage{
.heatmap_lines(
  loci,
  nbins,
  bin_size,
  upstream,
  downstream,
  mode,
  expand = TRUE
)
}
\arguments{
\item{loci}{Number of loci (rows)}

\item{nbins}{Number of bins (columns)}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{expand}{Whether to apply padding to the ticks.
Otherwise a strange box shows on heatmap.}
}
\value{
A list of ggproto objects to be plotted.
}
\description{
Helper function for calculating guide lines and labels in heatmap
}
