% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{.intersect_bw_and_granges}
\alias{.intersect_bw_and_granges}
\title{Intersect a BigWig file over loci on a GRanges object}
\usage{
.intersect_bw_and_granges(
  bw,
  granges,
  npoints,
  ignore_strand = FALSE,
  default_na = NA_real_,
  scaling = "none"
)
}
\arguments{
\item{bw}{BigWigFile object (rtracklayer).}

\item{granges}{GRanges object.}

\item{npoints}{How many points to take (different to bin size!).}

\item{ignore_strand}{Ignore strand information in granges.}

\item{default_na}{Value to use as default for missing values}

\item{scaling}{none if no scaling is done, relative if scaled to global
mean of 1}
}
\value{
A value matrix of dimensions len(granges) x npoints.
}
\description{
Intersect a BigWig file over loci on a GRanges object, taking npoints points
per locus.
}
