% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_multiple_loci}
\alias{.is_multiple_loci}
\title{Check if a variable contains a list of things or array of potentially file names}
\usage{
.is_multiple_loci(loci)
}
\arguments{
\item{loci}{A list of GRanges / files / mixed (or not)}
}
\value{
TRUE if variable contains multiple loci, FALSE otherwise
}
\description{
Check if a variable contains a list of things or array of potentially file names
}
