% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.label_multiple_loci}
\alias{.label_multiple_loci}
\title{Label loci with a list of labels}
\usage{
.label_multiple_loci(loci, labels)
}
\arguments{
\item{loci}{List of loci}

\item{labels}{List of labels}
}
\value{
A list of labels
}
\description{
If some labels are duplicated, numbers are added to the end
}
