% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.limited_size_caption_line}
\alias{.limited_size_caption_line}
\title{Make a string out of a named list. Split into lines if too wide.}
\usage{
.limited_size_caption_line(named_list)
}
\arguments{
\item{named_list}{A named list}
}
\value{
A string
}
\description{
Make a string out of a named list. Split into lines if too wide.
}
