% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_caption}
\alias{.make_caption}
\title{Make a string to put as caption in verbose mode. Includes system date.}
\usage{
.make_caption(params, outcome)
}
\arguments{
\item{params}{Named list with relevant parameters and their values}

\item{outcome}{Named values with relevant outcomes and their values}
}
\value{
A caption string
}
\description{
Make a string to put as caption in verbose mode. Includes system date.
}
