% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.match_seqnames}
\alias{.match_seqnames}
\title{Match seqname lists and throws the relevant warning}
\usage{
.match_seqnames(seqnamelist, show_max_shared, show_max_missing)
}
\arguments{
\item{seqnamelist}{Names list to compare}

\item{show_max_shared}{Maximum shared references to print}

\item{show_max_missing}{Maximum missing references to print}
}
\value{
TRUE if they match, FALSE if they don't
}
\description{
Match seqname lists and throws the relevant warning
}
