% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{.multi_bw_ranges_aggregated}
\alias{.multi_bw_ranges_aggregated}
\title{Intersect a list of bigWig files with a GRanges object and aggregate by name}
\usage{
.multi_bw_ranges_aggregated(
  bwfiles,
  labels,
  granges,
  per_locus_stat,
  aggregate_by,
  remove_top,
  default_na = NA_real_,
  scaling = "none"
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{granges}{GRanges object to summarize. Should have a valid name field.}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{aggregate_by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{default_na}{Default value for missing values}

\item{scaling}{If none, no operation is performed (default). If relative,
values are divided by global mean (1x genome coverage).}
}
\value{
An aggregated dataframe
}
\description{
Intersect a list of bigWig files with a GRanges object and aggregate by name
}
