% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{.multi_bw_ranges_norm}
\alias{.multi_bw_ranges_norm}
\title{Intersect a list of bigWig files with a GRanges object (with background)}
\usage{
.multi_bw_ranges_norm(
  bwfilelist,
  bg_bwfilelist,
  labels,
  granges,
  per_locus_stat = "mean",
  selection = NULL,
  norm_func = identity,
  remove_top = 0,
  default_na = NA_real_
)
}
\arguments{
\item{bwfilelist}{BigWig file list to be summarized.}

\item{bg_bwfilelist}{Background BigWig files.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{granges}{GRanges object to intersect}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{selection}{A GRanges object to restrict analysis to.}

\item{norm_func}{Function to apply after bw/bg_bw.}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{default_na}{Default value for missing values}
}
\value{
a sorted GRanges object
}
\description{
Intersect a list of bigWig files with a GRanges object and normalize values
with a set of background bigWig files.
}
