% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.multi_ranges_overlap}
\alias{.multi_ranges_overlap}
\title{Calculate overlap of one GRanges with a main GRanges object}
\usage{
.multi_ranges_overlap(main_ranges, other_ranges, labels, minoverlap)
}
\arguments{
\item{main_ranges}{Main GRanges to which overlap is calculated.}

\item{other_ranges}{A list of GRanges objects}

\item{labels}{Labels to each of the other_ranges.}

\item{minoverlap}{Minimum overlap to consider an overlap.}
}
\value{
A data.frame in tall format with the values of the overlapping loci.
Loci returned belong to other_ranges, NOT main_ranges. A group field
is added as factor.
}
\description{
Calculate overlap of one GRanges with a main GRanges object
}
