% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.preprocess_heatmap_matrix}
\alias{.preprocess_heatmap_matrix}
\title{Preprocess heatmap values for plotting}
\usage{
.preprocess_heatmap_matrix(values, zmin, zmax, order_by, max_rows_allowed)
}
\arguments{
\item{values}{Matrix to preprocess}

\item{zmin}{Minimum value to show in color.}

\item{zmax}{Maximum value to show in color.}

\item{order_by}{Row order (or NULL).}

\item{max_rows_allowed}{Maximum heatmap resolution.}
}
\value{
A named list with values and stats calculated.
}
\description{
This includes capping on zmin / zmax values. If not provided, percentiles
0.01 and 0.99 are used, to match plotting of other similar tools.
If provided, a specific order of rows is used. Otherwise, rows are sorted
by rowMeans.
If the number of rows in the matrix is larger than a certain max_rows_allowed
value, number of rows is reduced to max_rows_allowed. Rather than subsampling
each row then will represent an average of the underlying rows.
}
