% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{.process_norm_mode}
\alias{.process_norm_mode}
\title{Get a function out of a norm mode string.}
\usage{
.process_norm_mode(mode)
}
\arguments{
\item{mode}{String representing mode. Valid values are fc and log2fc}
}
\value{
A function
}
\description{
Get a function out of a norm mode string.
}
