% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.profile_breaks}
\alias{.profile_breaks}
\title{Compute where the break ticks go in heatmap and profile}
\usage{
.profile_breaks(nrows, upstream, downstream, bin_size, mode)
}
\arguments{
\item{nrows}{Number of bins}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}
}
\value{
Array of numeric
}
\description{
Compute where the break ticks go in heatmap and profile
}
