% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.profile_labels}
\alias{.profile_labels}
\title{Make profile labels}
\usage{
.profile_labels(upstream, downstream, mode)
}
\arguments{
\item{upstream}{Basepairs upstream}

\item{downstream}{Basepairs downstream}

\item{mode}{Plot mode (stretch, start, end, center)}
}
\value{
Array of strings
}
\description{
Make profile labels
}
