% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.repeated_elements}
\alias{.repeated_elements}
\title{Test if an array has repeated element}
\usage{
.repeated_elements(array)
}
\arguments{
\item{array}{An array of char}
}
\value{
TRUE if all elements in the array are unique
}
\description{
Test if an array has repeated element
}
