% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.round_ignore_null}
\alias{.round_ignore_null}
\title{Round a value. If it's NULL, returns NULL.}
\usage{
.round_ignore_null(v, digits = 3)
}
\arguments{
\item{v}{Value}

\item{digits}{How many digits round}
}
\value{
Rounded value, or NULL
}
\description{
Round a value. If it's NULL, returns NULL.
}
