% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{.summary_body}
\alias{.summary_body}
\title{Helper function for plotting a summary matrix}
\usage{
.summary_body(values)
}
\arguments{
\item{values}{Summary matrix}
}
\description{
Helper function for plotting a summary matrix
}
