% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.trunc_str}
\alias{.trunc_str}
\title{Truncate a string to a max length}
\usage{
.trunc_str(s, max_length = 35)
}
\arguments{
\item{s}{String to truncate}

\item{max_length}{Length to truncate s to (35)}
}
\description{
Truncate a string to a max length
}
