% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_categories}
\alias{.validate_categories}
\title{Validate a category array}
\usage{
.validate_categories(cat_values)
}
\arguments{
\item{cat_values}{An array of values}
}
\description{
Checks whether the number of categories is reasonable for an aggregation.
Throws a warning if it finds more than 50 different values.
}
