% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_filelist}
\alias{.validate_filelist}
\title{Validate an array of paths}
\usage{
.validate_filelist(filelist)
}
\arguments{
\item{filelist}{An array of files}
}
\value{
Nothing (NULL) on success.
}
\description{
Check that a list of files is valid: not empty and contents exist.
}
