% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_group_col}
\alias{.validate_group_col}
\title{Validate that group col exists in granges}
\usage{
.validate_group_col(granges, group_col)
}
\arguments{
\item{granges}{GRanges object to check}

\item{group_col}{Group column name. Usually, name.}
}
\value{
Nothing (NULL) on success.
}
\description{
Validate that group col exists in granges
}
