% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_locus_parameter}
\alias{.validate_locus_parameter}
\title{Validate that a locus parameter is valid. Checks for paths and also whether
the parameter is otherwise a GRanges object, which is also valid.}
\usage{
.validate_locus_parameter(locus_param)
}
\arguments{
\item{locus_param}{Parameter to validate}
}
\value{
Nothing (NULL) on success.
}
\description{
Validate that a locus parameter is valid. Checks for paths and also whether
the parameter is otherwise a GRanges object, which is also valid.
}
