% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_references}
\alias{.validate_references}
\title{Check that a list of bigWig files have the same Seqinfo.
Throws a warning if they don't.}
\usage{
.validate_references(bwfiles, show_max = 25)
}
\arguments{
\item{bwfiles}{List of bigWig files}

\item{show_max}{Max number of references to show in the warning}
}
\value{
TRUE if they share all references, FALSE otherwise.
}
\description{
Check that a list of bigWig files have the same Seqinfo.
Throws a warning if they don't.
}
