% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_references_with_external_seqinfo}
\alias{.validate_references_with_external_seqinfo}
\title{Check that a list of bigWig files have the same Seqinfo as an external object
(usually a pre-built tile set). Throws a warning if they don't.}
\usage{
.validate_references_with_external_seqinfo(bwfiles, seqinfo, show_max = 25)
}
\arguments{
\item{bwfiles}{List of bigWig files}

\item{seqinfo}{Seqinfo object}

\item{show_max}{Max number of shared references to show in the warning}
}
\value{
TRUE if they share all references, FALSE otherwise.
}
\description{
Check that a list of bigWig files have the same Seqinfo as an external object
(usually a pre-built tile set). Throws a warning if they don't.
}
