% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{estimate_read_counts}
\alias{estimate_read_counts}
\title{Estimate the number of reads in a locus}
\usage{
estimate_read_counts(mean_cov, width, fraglen)
}
\arguments{
\item{mean_cov}{Mean coverage}

\item{width}{Width of the locus}

\item{fraglen}{Fragment length}
}
\value{
An integer representing the estimated number of reads
(mean_cov*width) / fragment_length
}
\description{
This is a helper tool to back-calculate a number that represents the raw
counts in a locus. It is an estimate so one should consider several factors:
a) fraglen parameter needs to be accurate; b) fraglen is a constant, so if
the original distribution of fragment length was very heterogeneous, it
will over/under estimate locus where fragments are shorter or longer than
the average. c) If there was a scaling done in the original bigWig, this
number will not have a 1 to 1 correspondence with the original number of
reads.
}
