% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{fetch_bigwig}
\alias{fetch_bigwig}
\title{Make a BigWigFile object out of a path or an URL}
\usage{
fetch_bigwig(bw)
}
\arguments{
\item{bw}{BigWig file path or URL}
}
\value{
BigWigFile object
}
\description{
Make a BigWigFile object out of a path or an URL
}
