% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{get_heatmap_values}
\alias{get_heatmap_values}
\title{Get heatmap fixture from bw_heatmap for test plotting}
\usage{
get_heatmap_values()
}
\description{
Get heatmap fixture from bw_heatmap for test plotting
}
