% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{get_testfile}
\alias{get_testfile}
\title{Return the full path of a test file in wigglescout package}
\usage{
get_testfile(fname)
}
\arguments{
\item{fname}{Name of the file (basename)}
}
\value{
Full path of the file
}
\description{
Return the full path of a test file in wigglescout package
}
