% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{gr_coverage_to_read_counts}
\alias{gr_coverage_to_read_counts}
\title{Convert a GRanges object with coverage values to estimated counts}
\usage{
gr_coverage_to_read_counts(gr, mcol_names, fraglen)
}
\arguments{
\item{gr}{GRanges object}

\item{mcol_names}{Columns to calculate on}

\item{fraglen}{Fragment length to use for the estimate}
}
\value{
A GRanges object with count values.
}
\description{
This method uses estimate_read_counts across a GRanges object and converts
the values to read counts.
}
