% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{keep_canonical}
\alias{keep_canonical}
\title{Filter out unplaced contigs and uncharacterized chromosomes from a GRanges
object.}
\usage{
keep_canonical(gr)
}
\arguments{
\item{gr}{A filtered GRanges object}
}
\value{
A filtered GRanges object
}
\description{
This utility will keep only the sequences that belong to canonical chromosomes
returned by GenomeInfoDb::standardChromosomes function.
}
\examples{
bins_gr <- build_bins(bin_size = 50000, genome = "hg38")
keep_canonical(bins_gr)
}
