% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{key_value_string}
\alias{key_value_string}
\title{Make a string out of a named list.}
\usage{
key_value_string(named_list)
}
\arguments{
\item{named_list}{A named list}
}
\value{
A string
}
\description{
Make a string out of a named list.
}
