% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{local_create_sample_bigwig}
\alias{local_create_sample_bigwig}
\title{Create a self-destructing local temporary bigwig based on a BED file and
chromsizes array.}
\usage{
local_create_sample_bigwig(
  bed,
  chromsizes,
  f = tempfile(fileext = ".bw"),
  env = parent.frame()
)
}
\arguments{
\item{bed}{BED filename}

\item{chromsizes}{Vector with length of chromosomes}

\item{f}{bigWig filename}

\item{env}{Running environment}
}
\value{
Local filename
}
\description{
Create a self-destructing local temporary bigwig based on a BED file and
chromsizes array.
}
