% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{make_caption}
\alias{make_caption}
\title{Make a string to put as capsion in verbose mode. Includes system date.}
\usage{
make_caption(params, outcome)
}
\arguments{
\item{params}{Named list with relevant parameters and their values}

\item{outcome}{Named values with relevant outcomes and their values}
}
\value{
A caption string
}
\description{
Make a string to put as capsion in verbose mode. Includes system date.
}
