% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_label_from_filename}
\alias{make_label_from_filename}
\title{Get a valid label from a filename}
\usage{
make_label_from_filename(filename)
}
\arguments{
\item{filename}{File to convert to label}
}
\value{
A valid label name
}
\description{
Get a valid label from a filename
}
