% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{make_mock_bins_all}
\alias{make_mock_bins_all}
\title{Make a mock object that returns bins GRanges for test plotting}
\usage{
make_mock_bins_all()
}
\value{
GRanges object
}
\description{
Make a mock object that returns bins GRanges for test plotting
}
