% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{make_mock_profile}
\alias{make_mock_profile}
\title{Make a mock object from a fixture that has a profile result from bw_profile
for test plotting}
\usage{
make_mock_profile()
}
\description{
Make a mock object from a fixture that has a profile result from bw_profile
for test plotting
}
