% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{make_mock_summary}
\alias{make_mock_summary}
\title{Make a mock object from a fixture that has a summary result from bw_heatmap
for test plotting}
\usage{
make_mock_summary()
}
\description{
Make a mock object from a fixture that has a summary result from bw_heatmap
for test plotting
}
