% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_norm_label}
\alias{make_norm_label}
\title{Generate a human-readable normalization function string}
\usage{
make_norm_label(f, bg)
}
\arguments{
\item{f}{String representing normalization function.}

\item{bg}{Background file.}
}
\value{
A string describing normalization.
}
\description{
Generate a human-readable normalization function string
}
