% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{multi_bw_ranges}
\alias{multi_bw_ranges}
\title{Intersect a list of bigWig files with a GRanges object}
\usage{
multi_bw_ranges(
  bwfiles,
  labels,
  granges,
  per_locus_stat = "mean",
  selection = NULL,
  remove_top = 0
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{granges}{GRanges object to intersect}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}
}
\value{
A sorted GRanges object.
}
\description{
Build a binned-scored GRanges object from a list of bigWig files. The
aggregating function per locus can be min, max, sd, mean.
}
