% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{natural_sort_by_field}
\alias{natural_sort_by_field}
\title{Take a column of a data frame to use as row names, drop such column and
reorder the data frame so row names follow natural order.}
\usage{
natural_sort_by_field(df, col)
}
\arguments{
\item{df}{A data frame.}

\item{col}{The column (usually name).}
}
\value{
A sorted df
}
\description{
Take a column of a data frame to use as row names, drop such column and
reorder the data frame so row names follow natural order.
}
