% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bins_density}
\alias{plot_bw_bins_density}
\title{Bin-based 2d histogram of a pair of bigWig files}
\usage{
plot_bw_bins_density(
  x,
  y,
  bg_x = NULL,
  bg_y = NULL,
  norm_mode_x = "fc",
  norm_mode_y = "fc",
  bin_size = 10000,
  genome = "mm9",
  plot_binwidth = 0.05,
  highlight = NULL,
  remove_top = 0,
  verbose = TRUE,
  selection = NULL,
  default_na = NA_real_,
  scaling = "none",
  canonical = FALSE
)
}
\arguments{
\item{x}{BigWig file corresponding to the x axis.}

\item{y}{BigWig file corresponding to the y axis.}

\item{bg_x}{BigWig file to be used as x axis background (us. input).}

\item{bg_y}{BigWig file to be used as y axis background (us. input).}

\item{norm_mode_x}{Normalization mode for x axis.}

\item{norm_mode_y}{Normalization mode for y axis.}

\item{bin_size}{Bin size.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{plot_binwidth}{Resolution of the bins in the density histogram
(different to genomic bin size)}

\item{highlight}{List of bed files to use as highlight for subgroups.}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}

\item{default_na}{Default value for missing values}

\item{scaling}{If none, no operation is performed (default). If relative,
values are divided by global mean (1x genome coverage).}

\item{canonical}{Use only canonical chromosomes (default: FALSE)}
}
\value{
A ggplot object.
}
\description{
Plots a 2d  binned histogram plot from two given bigWig files
}
\details{
Values in x and y axis can be normalized using background bigWig files
(usually input files). By default, the value shown will be x / bg_x per bin.
If norm_func_x or norm_func_y are provided, this can be changed to any given
function, for instance, if norm_func_x = log2, values on the x axis will
represent log2(x / bg_x) for each bin.

Values that are invalid (NaN, Inf, -Inf) in doing such normalization will
be ignored and shown as warnings, as this is ggplot default behavior.
}
\examples{
# Get the raw files
bw <- system.file("extdata", "sample_H33_ChIP.bw", package="wigglescout")
bw2 <- system.file("extdata",
                   "sample_H3K9me3_ChIP.bw", package="wigglescout")
bed <- system.file("extdata", "sample_genes_mm9.bed", package="wigglescout")

# Sample bigWig files only have valid values on this region
locus <- GenomicRanges::GRanges(
  seqnames = "chr15",
  IRanges::IRanges(102600000, 103100000)
)

plot_bw_bins_density(bw, bw2, bin_size = 50000, selection = locus)
plot_bw_bins_density(bw, bw2, bin_size = 50000, selection = locus,
    remove_top = 0.01)
}
