% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bins_scatter}
\alias{plot_bw_bins_scatter}
\title{Bin-based scatterplot of a pair of bigWig files}
\usage{
plot_bw_bins_scatter(
  x,
  y,
  bg_x = NULL,
  bg_y = NULL,
  norm_mode_x = "fc",
  norm_mode_y = "fc",
  bin_size = 10000,
  genome = "mm9",
  highlight = NULL,
  minoverlap = 0L,
  highlight_label = NULL,
  highlight_colors = NULL,
  remove_top = 0,
  verbose = TRUE
)
}
\arguments{
\item{x}{BigWig file corresponding to the x axis.}

\item{y}{BigWig file corresponding to the y axis.}

\item{bg_x}{BigWig file to be used as x axis background (us. input).}

\item{bg_y}{BigWig file to be used as y axis background (us. input).}

\item{norm_mode_x}{Normalization mode for x axis.}

\item{norm_mode_y}{Normalization mode for y axis.}

\item{bin_size}{Bin size.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{highlight}{List of bed files to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted}

\item{highlight_label}{Labels for the highlight groups.
If not provided, filenames are used.}

\item{highlight_colors}{Array of color values for the highlighting groups}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}
}
\value{
A ggplot object.
}
\description{
Plots a scatter plot from two given bigWig files and an optional set of BED
files as highlighted annotations. Bins are highlighted if there is at least
minoverlap base pairs overlap with any loci in BED file.
}
\details{
If specifying minoverlap, you must take into account the bin_size parameter
and the size of the loci you are providing as BED file.

Values in x and y axis can be normalized using background bigWig files
(usually input files). By default, the value shown will be x / bg_x per bin.
If norm_func_x or norm_func_y are provided, this can be changed to any given
function, for instance, if norm_func_x = log2, values on the x axis will
represent log2(x / bg_x) for each bin.

Values that are invalid (NaN, Inf, -Inf) in doing such normalization will
be ignored and shown as warnings, as this is ggplot default behavior.
}
