% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_bins_violin}
\alias{plot_bw_bins_violin}
\title{Bin-based violin plot of a set of bigWig files}
\usage{
plot_bw_bins_violin(
  bwfiles,
  bg_bwfiles = NULL,
  labels = NULL,
  bin_size = 10000,
  per_locus_stat = "mean",
  genome = "mm9",
  highlight = NULL,
  minoverlap = 0L,
  norm_mode = "fc",
  highlight_label = NULL,
  highlight_colors = NULL,
  remove_top = 0,
  verbose = TRUE,
  selection = NULL
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{bin_size}{Bin size.}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{highlight}{BED file to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{highlight_label}{Label for the highlighted loci set}

\item{highlight_colors}{Array of color values for the highlighted groups.}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}
}
\value{
A ggplot object.
}
\description{
Plots a violin plot of bin distribution of a set of bigWig files optionally
overlaid with annotated bins. Bins overlapping loci of the provided BED
file will be shown as a jitter plot on top of the violin plot.
}
