% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_heatmap}
\alias{plot_bw_heatmap}
\title{Plot a heatmap of a given bigWig file over a set of loci}
\usage{
plot_bw_heatmap(
  bwfile,
  bedfile,
  bg_bwfile = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  norm_mode = "fc",
  cmap = "Reds",
  zmin = NULL,
  zmax = NULL,
  max_rows_allowed = 10000,
  order_by = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bwfile}{BigWig file to plot}

\item{bedfile}{BED file to summarize}

\item{bg_bwfile}{Background bw file. Use this with care. Depending on bin
size and actual values, this may result in a very noisy plot.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{middle}{Number of base pairs that the middle section has (in stretch
mode). If not provided, median length of all loci is used.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{cmap}{Color map. Any RColorBrewer palette name is accepted here.}

\item{zmin}{Minimum of the color scale. Majority of tools set
this to 0.01 percentile of the data distribution.}

\item{zmax}{Maximum of the color scale. Majority of tools set
this to 0.99.}

\item{max_rows_allowed}{Maximum number of loci that will be allowed in the
plot. If the amount of loci exceeds this value, the plot will be binned
on the y axis until it fits max_rows_allowed. This speeds up plotting of
very large matrices, where higher resolution would not be perceivable by eye.}

\item{order_by}{Specific order to display rows. By default rows are sorted
decreasingly by mean. Order should be an array of integers of the same
length as number of rows. These can be obtained as a result of order()
function, if one would want to sort one heatmap by values on another one.}

\item{verbose}{Put a caption with relevant parameters on the plot.}
}
\description{
Plot a heatmap of a given bigWig file over a set of loci
}
