% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_profile}
\alias{plot_bw_profile}
\title{Profile plot of a set of bigWig files}
\usage{
plot_bw_profile(
  bwfiles,
  loci,
  bg_bwfiles = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  show_error = FALSE,
  norm_mode = "fc",
  labels = NULL,
  colors = NULL,
  remove_top = 0,
  verbose = TRUE
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{loci}{BED file or GRanges to summarize}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{middle}{Number of base pairs that the middle section has (in stretch
mode). If not provided, median length of all loci is used.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{show_error}{Show standard error.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{colors}{Array of colors that will  be assigned to labels or files
(in that order)}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}
}
\value{
A ggplot object.
}
\description{
Plots a profile of a set of bigWig files over a set of loci in a BED file.
}
\examples{
# Get the raw files
bw <- system.file("extdata", "sample_H33_ChIP.bw", package="wigglescout")
bed <- system.file("extdata", "sample_genes_mm9.bed", package="wigglescout")


plot_bw_profile(bw, loci = bed,
                mode = "stretch", upstream = 1000, downstream = 1000)

# It is also possible to run same bigWig across several BED files or GRanges
plot_bw_profile(bw, loci = c(bed, bed), labels = c("A", "B"))
}
