% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_gr_scatter}
\alias{plot_gr_scatter}
\title{Bin-based scatterplot of a GRanges object}
\usage{
plot_gr_scatter(
  gr,
  x,
  y,
  highlight = NULL,
  minoverlap = 0L,
  highlight_label = NULL,
  highlight_colors = NULL,
  remove_top = 0,
  verbose = TRUE,
  selection = NULL
)
}
\arguments{
\item{gr}{Scored GRanges object}

\item{x}{Column in gr corresponding to the x axis}

\item{y}{Column in gr corresponding to the y axis.}

\item{highlight}{List of bed files to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted}

\item{highlight_label}{Labels for the highlight groups.
If not provided, filenames are used.}

\item{highlight_colors}{Array of color values for the highlighting groups}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}
}
\value{
A ggplot object.
}
\description{
Plots a scatter plot from two mcols of a GRanges object files and an optional
set of BED files as highlighted annotations. Bins are highlighted if there is
at least minoverlap base pairs overlap with any loci in BED file.
}
\details{
If specifying minoverlap, you must take into account the bin_size parameter
and the size of the loci you are providing as BED file.

This function does not calculate background normalization or anything, you
can do that in the prior call to bw_bins or bw_loci
}
