% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_gr_violin}
\alias{plot_gr_violin}
\title{Violin plot of a precalculated GRanges object}
\usage{
plot_gr_violin(
  gr,
  columns,
  highlight = NULL,
  minoverlap = 0L,
  highlight_label = NULL,
  highlight_colors = NULL,
  remove_top = 0,
  verbose = TRUE,
  selection = NULL
)
}
\arguments{
\item{gr}{Scored GRanges object}

\item{columns}{Columns in gr to plot}

\item{highlight}{BED file to use as highlight for subgroups.}

\item{minoverlap}{Minimum overlap required for a bin to be highlighted.}

\item{highlight_label}{Label for the highlighted loci set}

\item{highlight_colors}{Array of color values for the highlighted groups.}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}
}
\value{
A ggplot object.
}
\description{
Plots a violin plot of bin distribution of a set of bigWig files optionally
overlaid with annotated bins. Bins overlapping loci of the provided BED
file will be shown as a jitter plot on top of the violin plot.
}
