% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_matrix_profile}
\alias{plot_matrix_profile}
\title{Helper function plots a profile from a dataframe}
\usage{
plot_matrix_profile(values, show_error, colors)
}
\arguments{
\item{values}{Values data frame in long format}

\item{show_error}{Boolean wheter tho show error.}

\item{colors}{Alternative colors to plot the lines}
}
\value{
A ggplot object
}
\description{
Helper function plots a profile from a dataframe
}
