% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_matrix_summary}
\alias{plot_matrix_summary}
\title{Helper function for plotting a summary matrix}
\usage{
plot_matrix_summary(values)
}
\arguments{
\item{values}{Summary matrix}
}
\description{
Helper function for plotting a summary matrix
}
