% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{process_highlight_loci}
\alias{process_highlight_loci}
\title{Internal processing of loci sets and labels.}
\usage{
process_highlight_loci(loci_sets, labels)
}
\arguments{
\item{loci_sets}{One or more BED files or GRanges. These need to be either
all BED files or all GRanges objects.}

\item{labels}{One or more labels. Lengths need to match loci_sets or be NULL,
in which case loci_sets are assumed to be BED files.}
}
\value{
a named list with ranges and labels.
}
\description{
This allows to process equally lists of GRanges objects or BED files. Note
that GRanges objects need to be lists, as c(GRanges) concatenates elements
in single GRanges object
}
