% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{remove_top_by_mean}
\alias{remove_top_by_mean}
\title{Remove top loci in a GRanges column by mean of specified values}
\usage{
remove_top_by_mean(granges, quantile, columns)
}
\arguments{
\item{granges}{GRanges object. Must have numerical mcols, and names of those
needs to match with columns.}

\item{quantile}{Top quantile. Must be a number between zero and 1.}

\item{columns}{Which columns to use. If more than one, quantile will be
selected according to means.}
}
\value{
A named list, fields ranges for the resulting GRanges, plus calculated
values: quantile_value, filtered, na_values.
}
\description{
Remove top loci in a GRanges column by mean of specified values
}
