% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{tile_seqinfo}
\alias{tile_seqinfo}
\title{Build a tiles GRanges out of a bigWig file on a given bin size.}
\usage{
tile_seqinfo(seqinfo, bin_size)
}
\arguments{
\item{seqinfo}{A Seqinfo object}

\item{bin_size}{Size of the tile in basepairs}
}
\value{
A GRanges object with bins of bin_size bins matching the genome the
bigWig file was mapped to.
}
\description{
It does not need the reference genome as that information can be parsed from
the bigwig file
}
\examples{

tile_seqinfo(GenomeInfoDb::Seqinfo(genome = "mm9"), 50000)
}
