% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_bins}
\alias{bw_bins}
\title{Build a binned-scored GRanges object from a set of bigWig files}
\usage{
bw_bins(
  bwfiles,
  bg_bwfiles = NULL,
  labels = NULL,
  per_locus_stat = "mean",
  bin_size = 10000,
  genome = "mm9",
  selection = NULL,
  norm_mode = "fc",
  remove_top = 0,
  default_na = NA_real_
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{per_locus_stat}{Aggregating function (per locus). Mean by default.
Choices: min, max, sd, mean.}

\item{bin_size}{Bin size.}

\item{genome}{Genome. Available choices are mm9, hg38.}

\item{selection}{A GRanges object to restrict binning to a certain set of
intervals. This is useful for debugging and improving performance of
locus specific analyses.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{default_na}{Default value for missing values}
}
\value{
A GRanges object with each bwfile as a metadata column named
after labels, if provided, or after filenames otherwise.
}
\description{
Build a binned-scored GRanges object from one or many bigWig files.
The aggregating function per bin can be min, max, sd, mean.
}
\details{
bwfiles and bg_bwfiles must have the same length. If you are using same
background for several files, then file paths must be repeated accordingly.

norm_mode can be either "fc", where values will represent bigwig / bg_bigwig,
or "log2fc", values will represent log2(bigwig / bg_bigwig) per bin.
}
\examples{
# Get the raw files
bw <- system.file("extdata", "sample_H33_ChIP.bw", package="wigglescout")
bw2 <- system.file("extdata",
                   "sample_H3K9me3_ChIP.bw", package="wigglescout")
bw_inp <- system.file("extdata", "sample_Input.bw", package="wigglescout")

# Sample bigWig files only have valid values on this region
locus <- GenomicRanges::GRanges(
  seqnames = "chr15",
  IRanges::IRanges(102600000, 103100000)
)

# Run single bw. The larger the bin size, the faster this is.
bw_bins(bw, bin_size = 100000, labels = c("H33"), selection = locus)

# Use of some parameters
bw_bins(bw, bg_bwfiles = bw_inp, bin_size = 100000,
       labels = c("H33"), norm_mode = "log2fc", selection = locus)

# Multiple bigWig
bw_bins(c(bw, bw2),
        bin_size = 100000,
        bg_bwfiles = c(bw_inp, bw_inp),
        norm_mode = "log2fc",
        selection = locus)
}
