% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtools.R
\name{bw_heatmap}
\alias{bw_heatmap}
\title{Calculate heatmap matrix of a bigWig file over a GRanges or BED file}
\usage{
bw_heatmap(
  bwfiles,
  bg_bwfiles = NULL,
  loci = NULL,
  labels = NULL,
  mode = "stretch",
  bin_size = 100,
  upstream = 2500,
  downstream = 2500,
  middle = NULL,
  ignore_strand = FALSE,
  norm_mode = "fc",
  default_na = NA_real_,
  scaling = "none"
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{loci}{BED file or GRanges to summarize}

\item{labels}{List of names to give to the mcols of the returned GRanges
object. If NULL, file names are used.}

\item{mode}{How to handle differences in lengths across loci:

stretch: Anchor each locus on both sides.

start: Anchor all loci on start.

end: Anchor all loci on end.

center: Center all loci.}

\item{bin_size}{Bin size. Length of bin in base pairs. The lower, the higher
the resolution.}

\item{upstream}{Number of base pairs to include upstream of loci.}

\item{downstream}{Number of base pairs to include downstream of loci.}

\item{middle}{Number of base pairs that the middle section has (in stretch
mode). If not provided, median length of all loci is used.}

\item{ignore_strand}{Whether to use strand information in BED file.}

\item{norm_mode}{Function to apply to normalize bin values. Default fc:
divides bw / bg. Alternative: log2fc: returns log2(bw/bg).}

\item{default_na}{Default value for missing values}

\item{scaling}{Whether to use the bigWig values as they are (none - default)
or calculate relative enrichment (relative) by dividing values by global
average.}
}
\value{
A list of matrices where each element correspond to each bigWig file.
}
\description{
Calculates a matrix of values that corresponds to the usual heatmaps where
each row is a locus and columns are data points taken each bin_size base
pairs.
}
\details{
Loci are aligned depending on mode parameter:
\itemize{
\item stretch. Aligns all starts and all ends, sort of stretching the loci.
The median of these lenghts is taken as the pseudo-length in order to show
a realistic plot when displayed.
\item start. All loci are aligned by start.
\item end. All loci are aligned by end.
\item center. All loci are aligned by center.
}
}
\examples{
# Get the raw files
bed <- system.file("extdata", "sample_genes_mm9.bed", package="wigglescout")
bw <- system.file("extdata", "sample_H33_ChIP.bw", package="wigglescout")
bw2 <- system.file("extdata",
                   "sample_H3K9me3_ChIP.bw", package="wigglescout")

# Heatmaps with a single bigWig
h <- bw_heatmap(bw, loci = bed, mode = "stretch")

# h is a list
h[[1]]

# Heatmaps with multiple bigWig
h <- bw_heatmap(c(bw, bw2), loci = bed, mode = "stretch")

# h is a list
h[[2]]

}
