% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.granges_cbind}
\alias{.granges_cbind}
\title{GRanges left join operation}
\usage{
.granges_cbind(grlist, labels, granges = NULL)
}
\arguments{
\item{grlist}{A list of GRanges objects that have all the same fields.}

\item{labels}{Vector of names for the score columns.}

\item{granges}{Optional granges with name field on it}
}
\value{
A Sorted GRanges object with all the columns.
}
\description{
Perform a left_bind operation on a GRanges list, appending scores to mcols.
}
\details{
This assumes that you are trying to join things that match (bins generated
from the same parameters, BED intersections from the same BED.)
}
