% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.loci_to_granges}
\alias{.loci_to_granges}
\title{Processes a loci and returns a GRanges object, sorted and with its seqlevels
also sorted.}
\usage{
.loci_to_granges(loci)
}
\arguments{
\item{loci}{Either a BED file or a GRanges object}
}
\value{
A GRanges object
}
\description{
Processes a loci and returns a GRanges object, sorted and with its seqlevels
also sorted.
}
