% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwplot.R
\name{plot_bw_loci_summary_heatmap}
\alias{plot_bw_loci_summary_heatmap}
\title{Summary heatmap of a categorized BED or GRanges object}
\usage{
plot_bw_loci_summary_heatmap(
  bwfiles,
  loci,
  bg_bwfiles = NULL,
  labels = NULL,
  aggregate_by = "true_mean",
  norm_func = identity,
  remove_top = 0,
  verbose = TRUE
)
}
\arguments{
\item{bwfiles}{Path or array of paths to the bigWig files to be summarized.}

\item{loci}{BED file or GRanges object.}

\item{bg_bwfiles}{Path or array of paths to the bigWig files to be used as
background.}

\item{labels}{Labels to use for in the plot for the bw files.}

\item{aggregate_by}{Statistic to aggregate per group. If NULL, values are
not aggregated. This is the behavior by default.}

\item{norm_func}{Function to apply to normalize bin values f(bw / bw_bg).}

\item{remove_top}{Return range 0-(1-remove_top). By default returns the
whole distribution (remove_top == 0).}

\item{verbose}{Put a caption with relevant parameters on the plot.}
}
\value{
A ggplot object
}
\description{
Make a summary heatmap where each cell contains an aggregated value of a
bigWig file from bwfiles and a category of a BED file or GRanges (loci). The
provided loci must have a name field that is valid (i.e. can be grouped,
representing some type of category).
}
